Dashboard api and employee Conversion

+++ AttendanceSystem/ViewModel/WebAPI/DashboardCountVM.cs
@@ -5,7 +5,7 @@
         public int TotalAttendance { get; set; }
         public int TotalAbsent { get; set; }
         public int LeavePendingForApprove { get; set; }
-        public long PendingSalary { get; set; }
+        public decimal PendingSalary { get; set; }
         public double thisMonthHoliday { get; set; }
         public string LastMonthRating { get; set; }
         public int AttendancePendingForApprove { get; set; }
